Improve `Place.label` rendering
We used the OSM `display_name` to show the address.
The issue is that values of this attribute are quite complex.
It seems to be the full combination of all elements in `address` details.

In everyday usages, we do not use all of these details.
So we adapt the label to be more user friendly.

This commit does the following:

- Specify the address detail in model
- Create a helper to have a more user friendly address label